feed/feed/requirements.txt
requests
stream-python